requirements.txt
# ── Core ──────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0

# ── Machine Learning ──────────────────────────────────
scikit-image>=0.21.0        # Feature extraction (HOG, LBP)
joblib>=1.3.0               # Model saving/loading

# ── Deep Learning ─────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0

# ── Generative AI ─────────────────────────────────────
transformers>=4.35.0        # HuggingFace: BLIP, ViT-GPT2
accelerate>=0.24.0          # Fast model loading
anthropic>=0.20.0           # Anthropic API (Claude vision)

# ── Utilities ─────────────────────────────────────────
python-dotenv>=1.0.0        # Load API keys from .env
rich>=13.0.0                # Pretty terminal output
